+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - Improvement: Return `methodOverride` middleware to http config;
 - Improvement: Add hidden `restPrefix` config to `blueprints` configuration file;
 - Improvement: Add hidden `serverOptions` and `bodyParser` configurations to `http` configuration file;
+- Improvement: Add `.gitkeep` files to each `api` folder;
+- Improvement: Extend `api` folder with other directories which loads with Sails by default;
 - Improvement: Optimizing and improvement yeoman `index.js` file structure;
 
 ## Version 0.4.0